Prettier

+++ contracts/test/ContractMetaMock.sol
@@ -31,7 +31,7 @@ contract ContractMetaMock {
     function _bytes32ToString(bytes32 b) internal pure returns (string memory s) {
         s = new string(32);
         uint256 len = 32;
-        for (uint i = 0; i < 32; ++i) {
+        for (uint256 i = 0; i < 32; ++i) {
             if (uint8(b[i]) == 0) {
                 len = i;
                 break;
